# Conflicts:
#   WattApp/server/Server/Controllers/UsersController.cs

+++ WattApp/client/src/app/components/admin/admin-dso/admin-dso.component.ts
@@ -9,14 +9,17 @@ import { AuthService } from 'src/app/services/auth.service';
   styleUrls: ['./admin-dso.component.css']
 })
 export class AdminDsoComponent {
-
+	currentPage:number=1;
+	itemsPerPage:number=10;
+	totalItems:number=10;
   
   showUsers:ShowUsers[]=[];
   constructor(private router:Router,private usersService:AuthService,
     private route:ActivatedRoute) { }
 
   ngOnInit(): void {
-    this.usersService.getAllUsers().subscribe(users => {
+    this.usersService.getAllUsers(1).subscribe(users => {
+	 this.totalItems=users.numberOfPages*this.itemsPerPage;
      this.showUsers=users.data.map((u:any)=>({
        id: u.id,
        name: u.name,
@@ -27,23 +30,49 @@ export class AdminDsoComponent {
      } as ShowUsers));
     });
     }
+	pageChanged(pageNumber:number){
+		this.currentPage=pageNumber;
+		this.usersService.getAllUsers(pageNumber,this.itemsPerPage).subscribe(users => {
+			this.totalItems=users.numberOfPages*this.itemsPerPage;
+			this.showUsers=users.data.map((u:any)=>({
+			  id: u.id,
+			  name: u.name,
+			  username: u.username,
+			  block: u.blocked,
+			  email: u.email,
+			  role: u.role
+			} as ShowUsers));
+		   });
+	}
 
 
   blockUser(id: number) {
-    
-     
+    const block= document.getElementById('block-popup');
+    if(block!=null)
+    {
+      block.addEventListener('click', () => {
         this.usersService.blockUser(id).subscribe(() => {
           const userIndex = this.showUsers.findIndex(user => user.id === id);
           this.showUsers[userIndex].block = true;
         });
+      }
+      );
+    }
+      
   }
   unblockUser(id: number) {
-   
-      
+    const unblock= document.getElementById('unblock-popup');
+    if(unblock!=null)
+    {
+      unblock.addEventListener('click', () => {
         this.usersService.unblockUser(id).subscribe(() => {
           const userIndex = this.showUsers.findIndex(user => user.id === id);
           this.showUsers[userIndex].block = false;
         });
+      });
+
+      
+    }
   }
 
 
@@ -54,14 +83,21 @@ export class AdminDsoComponent {
 
   delete(id:number)
   {
+    
+      
+   
+    const deletePopup= document.getElementById('delete-admin-popup');
+    
+    if(deletePopup!=null)
+    {
+     deletePopup.addEventListener('click', () => {
       this.usersService.delete(id)
-      .subscribe({
-        next:()=>{
+      .subscribe(()=>{
           this.router.navigate(['dashboard']);
           location.reload();
-        }
       });
-    
+     });
+    }
   }
